+++ src/test/java/edu/illinois/library/cantaloupe/resource/iiif/v2/ImageResourceTest.java
@@ -179,19 +179,6 @@ public class ImageResourceTest extends ResourceTest {
     }
 
     @Test
-    public void testGETContentDispositionHeaderSetToInline() throws Exception {
-        URI uri = getHTTPURI("/" + IMAGE + "/full/full/0/color.jpg");
-        tester.testContentDispositionHeaderSetToInline(uri);
-    }
-
-    @Test
-    public void testGETContentDispositionHeaderSetToAttachment()
-            throws Exception {
-        URI uri = getHTTPURI("/" + IMAGE + "/full/full/0/color.jpg");
-        tester.testContentDispositionHeaderSetToAttachment(uri);
-    }
-
-    @Test
     public void testGETEndpointEnabled() {
         Configuration config = Configuration.getInstance();
         config.setProperty(Key.IIIF_1_ENDPOINT_ENABLED, true);